Squashed warning from VC8 for 64bit. Why is backend_stack defined as size_t when it holds a Cntxt? Changed both versions of mvc_export_chunk to take size_t args so that it is possible in principle to export large bats (the callers should be changed to provide size_t as well).

+++ sql/src/tools/prog.c
@@ -72,7 +72,7 @@ usage(char *prog)
 int
 main(int argc, char **av)
 {
-	int curlen = 0, maxlen = BUFSIZ*8;
+	size_t curlen = 0, maxlen = BUFSIZ*8;
 	char *prog = *av;
 	opt *set = NULL;
 	int setlen = 0, time = 0;
@@ -168,7 +168,7 @@ main(int argc, char **av)
 			break;
 		}
 		while ((line = fgets(buf+curlen, 1024, fp)) != NULL) {
-			int n = strlen(line);
+			size_t n = strlen(line);
             		curlen += n;
             		if (curlen+1024 > maxlen) {
                			maxlen += 8*BUFSIZ;
